Rename v2_to_v1_decorator.py to compat.py
