+++ src/snowflake/cli/_plugins/nativeapp/v2_conversions/compat.py
